[refactor] 다른 도메인도 팩토리를 생성하여 기존 서비스에서 레포지토리를 분리하는 작업
