+++ src/main/java/potatoes/server/travel/factory/TravelFactory.java
@@ -1,5 +1,7 @@
 package potatoes.server.travel.factory;
 
+import static potatoes.server.utils.error.ErrorCode.*;
+
 import java.util.List;
 
 import org.springframework.stereotype.Component;
@@ -11,13 +13,12 @@ import potatoes.server.chat.repository.ChatRepository;
 import potatoes.server.infra.s3.S3UtilsProvider;
 import potatoes.server.travel.entity.Travel;
 import potatoes.server.travel.entity.TravelPlan;
-import potatoes.server.travel.entity.TravelUser;
 import potatoes.server.travel.model.TravelModel;
 import potatoes.server.travel.model.TravelPlanModel;
 import potatoes.server.travel.repository.TravelPlanRepository;
 import potatoes.server.travel.repository.TravelRepository;
-import potatoes.server.travel.repository.TravelUserRepository;
 import potatoes.server.user.entity.User;
+import potatoes.server.utils.error.exception.WeGoException;
 
 @Component
 @RequiredArgsConstructor
@@ -25,7 +26,6 @@ public class TravelFactory {
 	private final S3UtilsProvider s3;
 	private final TravelRepository travelRepository;
 	private final TravelPlanRepository travelPlanRepository;
-	private final TravelUserRepository travelUserRepository;
 	private final ChatRepository chatRepository;
 
 	public Travel createTravel(TravelModel request, MultipartFile image) {
@@ -47,13 +47,12 @@ public class TravelFactory {
 		return TravelPlan.from(request, imageUrl);
 	}
 
-	public void createOrganizer(Travel travel, User user) {
-		TravelUser travelUser = TravelUser.createOrganizer(travel, user);
-		travelUserRepository.save(travelUser);
-	}
-
 	public void createChat(Travel travel, User host) {
 		Chat chat = Chat.createTravelChat(travel, host);
 		chatRepository.save(chat);
 	}
-}
+
+	public Travel findTravel(Long travelId) {
+		return travelRepository.findById(travelId)
+			.orElseThrow(() -> new WeGoException(TRAVEL_NOT_FOUND));
+	}}
