+++ src/main/java/potatoes/server/travel/service/TravelService.java
@@ -22,13 +22,14 @@ import potatoes.server.travel.dto.TravelPlanResponse;
 import potatoes.server.travel.entity.Travel;
 import potatoes.server.travel.entity.TravelUser;
 import potatoes.server.travel.factory.TravelFactory;
+import potatoes.server.travel.factory.TravelUserFactory;
 import potatoes.server.travel.model.TravelModel;
 import potatoes.server.travel.model.TravelPlanModel;
 import potatoes.server.travel.repository.TravelPlanRepository;
 import potatoes.server.travel.repository.TravelRepository;
 import potatoes.server.travel.repository.TravelUserRepository;
 import potatoes.server.user.entity.User;
-import potatoes.server.user.repository.UserRepository;
+import potatoes.server.user.factory.UserFactory;
 import potatoes.server.utils.constant.ParticipantRole;
 import potatoes.server.utils.error.exception.WeGoException;
 
@@ -38,16 +39,17 @@ import potatoes.server.utils.error.exception.WeGoException;
 @Service
 public class TravelService {
 
-	private final UserRepository userRepository;
 	private final TravelRepository travelRepository;
 	private final TravelPlanRepository travelPlanRepository;
 	private final TravelUserRepository travelUserRepository;
 	private final BookmarkRepository bookmarkRepository;
 	private final TravelFactory travelFactory;
+	private final TravelUserFactory travelUserFactory;
+	private final UserFactory userFactory;
 
 	@Transactional
 	public void createTravel(Long userId, CreateTravelRequest request) {
-		User user = findUser(userId);
+		User user = userFactory.findUser(userId);
 
 		TravelModel travelModel = CreateTravelRequest.toModel(request);
 		Travel travel = travelFactory.createTravel(travelModel, request.travelImage());
@@ -55,11 +57,11 @@ public class TravelService {
 		List<TravelPlanModel> travelPlan = DetailTravelRequest.toModels(request.detailTravel(), travel);
 		travelFactory.createTravelPlans(travelPlan);
 		travelFactory.createChat(travel, user);
-		travelFactory.createOrganizer(travel, user);
+		travelUserFactory.createOrganizer(travel, user);
 	}
 
 	public TravelDetailResponse getDetails(Long travelId, Optional<Long> userId) {
-		Travel travel = findTravel(travelId);
+		Travel travel = travelFactory.findTravel(travelId);
 
 		Boolean participationFlag = userId
 			.map(uid -> travelUserRepository.existsByTravelIdAndUserId(travelId, uid))
@@ -96,8 +98,8 @@ public class TravelService {
 
 	@Transactional
 	public void participateInTravel(Long travelId, Long userId) {
-		User user = findUser(userId);
-		Travel travel = findTravel(travelId);
+		User user = userFactory.findUser(userId);
+		Travel travel = travelFactory.findTravel(travelId);
 
 		boolean existsCheckFlag = travelUserRepository.existsByTravelIdAndUserId(travelId, userId);
 
@@ -116,7 +118,7 @@ public class TravelService {
 
 	@Transactional
 	public void deleteTravelByOrganizer(Long travelId, Long userId) {
-		TravelUser travelUser = findTravelUser(travelId, userId);
+		TravelUser travelUser = travelUserFactory.findTravelUser(travelId, userId);
 
 		if (travelUser.getRole() != ParticipantRole.ORGANIZER) {
 			throw new WeGoException(INSUFFICIENT_TRAVEL_PERMISSION);
@@ -128,7 +130,7 @@ public class TravelService {
 
 	@Transactional
 	public void deleteTravelByAttendee(Long travelId, Long userId) {
-		TravelUser travelUser = findTravelUser(travelId, userId);
+		TravelUser travelUser = travelUserFactory.findTravelUser(travelId, userId);
 
 		if (travelUser.getRole() != ParticipantRole.ATTENDEE) {
 			throw new WeGoException(NOT_PARTICIPATED_TRAVEL);
@@ -139,8 +141,8 @@ public class TravelService {
 
 	@Transactional
 	public void addBookmark(Long userId, Long travelId) {
-		User user = findUser(userId);
-		Travel travel = findTravel(travelId);
+		User user = userFactory.findUser(userId);
+		Travel travel = travelFactory.findTravel(travelId);
 
 		if (bookmarkRepository.existsByUserAndTravel(user, travel)) {
 			return;
@@ -155,26 +157,11 @@ public class TravelService {
 
 	@Transactional
 	public void deleteBookmark(Long userId, Long travelId) {
-		User user = findUser(userId);
-		Travel travel = findTravel(travelId);
+		User user = userFactory.findUser(userId);
+		Travel travel = travelFactory.findTravel(travelId);
 
 		Bookmark bookmark = bookmarkRepository.findByUserAndTravel(user, travel)
 			.orElseThrow(() -> new WeGoException(BOOKMARK_NOT_FOUND));
 		bookmarkRepository.delete(bookmark);
-	}
-
-	private User findUser(Long userId) {
-		return userRepository.findById(userId)
-			.orElseThrow(() -> new WeGoException(USER_NOT_FOUND));
-	}
-
-	private Travel findTravel(Long travelId) {
-		return travelRepository.findById(travelId)
-			.orElseThrow(() -> new WeGoException(TRAVEL_NOT_FOUND));
-	}
-
-	private TravelUser findTravelUser(Long travelId, Long userId) {
-		return travelUserRepository.findByTravelIdAndUserId(travelId, userId)
-			.orElseThrow(() -> new WeGoException(TRAVEL_NOT_FOUND));
 	}
 }
